Only starting a RabbitMQ container if it doesnt point to an external IP address
