+++ vantage6-server/vantage6/server/__init__.py
@@ -24,7 +24,6 @@ from flask_socketio import SocketIO
 
 from vantage6.server import db
 from vantage6.cli.context import ServerContext
-from vantage6.cli.rabbitmq.queue_manager import split_rabbitmq_uri
 from vantage6.server.model.base import DatabaseSessionManager, Database
 from vantage6.server.resource._schema import HATEOASModelSchema
 from vantage6.common import logger_name
