+++ vantage6/vantage6/cli/rabbitmq/queue_manager.py
@@ -16,37 +16,11 @@ from vantage6.common.docker.network_manager import NetworkManager
 from vantage6.cli.context import ServerContext
 from vantage6.cli.rabbitmq.definitions import RABBITMQ_DEFINITIONS
 from vantage6.cli.globals import RABBIT_TIMEOUT
+from vantage6.cli.rabbitmq import split_rabbitmq_uri
 
 DEFAULT_RABBIT_IMAGE = 'harbor2.vantage6.ai/infrastructure/rabbitmq'
 RABBIT_CONFIG = 'rabbitmq.config'
 RABBIT_DIR = 'rabbitmq'
-
-
-def split_rabbitmq_uri(rabbit_uri: str) -> Dict:
-    """
-    Get details (user, pass, host, vhost, port) from a RabbitMQ uri
-
-    Parameters
-    ----------
-    rabbit_uri: str
-        URI of RabbitMQ service ('amqp://$user:$pass@$host:$port/$vhost')
-
-    Returns
-    -------
-    Dict[str]
-        The vhost defined in the RabbitMQ URI
-    """
-    (user_details, location_details) = rabbit_uri.split('@', 1)
-    (user, password) = user_details.split('/')[-1].split(':', 1)
-    (host, remainder) = location_details.split(':', 1)
-    port, vhost = remainder.split('/', 1)
-    return {
-        'user': user,
-        'password': password,
-        'host': host,
-        'port': port,
-        'vhost': vhost,
-    }
 
 
 class RabbitMQManager:
